ChampVA: Add specs for model 10_7959C and 10_7959C_REV2025
